+++ rita_client/src/error.rs
@@ -1,5 +1,5 @@
 use althea_kernel_interface::KernelInterfaceError;
-use althea_types::exit_encryption::ExitEncryptionError;
+use althea_types::ExitEncryptionError;
 use awc::error::{JsonPayloadError, SendRequestError};
 use babel_monitor::structs::BabelMonitorError;
 use compressed_log::builder::LoggerError;
