+++ rita_client/src/exit_manager/requests.rs
@@ -3,9 +3,9 @@ use super::DEFAULT_WG_LISTEN_PORT;
 use crate::rita_loop::CLIENT_LOOP_TIMEOUT;
 use crate::RitaClientError;
 use actix_web_async::Result;
-use althea_types::exit_encryption::decrypt_exit_list;
-use althea_types::exit_encryption::decrypt_exit_state;
-use althea_types::exit_encryption::encrypt_exit_client_id;
+use althea_types::decrypt_exit_list;
+use althea_types::decrypt_exit_state;
+use althea_types::encrypt_exit_client_id;
 use althea_types::ExitListV2;
 use althea_types::WgKey;
 use althea_types::{ExitClientIdentity, ExitRegistrationDetails, ExitState};
